requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
reportlab==4.0.4
twilio==8.10.0
requests==2.31.0
Pillow==10.0.1